+++ src/BinTracker.Services/Services.cs
@@ -502,6 +502,7 @@ public static class ServiceSetup
     public static IServiceCollection AddBinTrackerServices(this IServiceCollection services)
     {
         services.AddSingleton<UserSession>();
+        services.AddSingleton<ApplicationState>();
         services.AddScoped<IAuditService, AuditService>();
         services.AddScoped<IAuthenticationService, AuthenticationService>();
         services.AddScoped<IUserService, UserService>();